services/market-feed/requirements.txt
apscheduler==3.10.4
psycopg[binary]==3.2.3
pykrx==1.0.45
yfinance==0.2.50
pandas==2.2.3
python-dotenv==1.0.1
structlog==24.4.0
